# src/app.js
const express = require('express');
const path = require('path');
const swaggerUi = require('swagger-ui-express');
const YAML = require('yamljs');

const config = require('./config');
const requireHfToken = require('./middleware/requireHfToken');
const authRouter = require('./routes/auth');
const componentsRouter = require('./routes/components');
const dataRouter = require('./routes/data');
const realtimeRouter = require('./routes/realtime');
const huggingfaceCompletion = require('./routes/huggingface');
const hfSpaceProxy = require('./routes/hfSpaceProxy');
const apiV1Router = require('./routes/api/v1');

const DEFAULT_CITY = config.defaultCity;
const DEFAULT_SIGN = config.defaultSign;

const app = express();
const swaggerDocument = YAML.load(path.join(__dirname, '..', 'cimeika-api.yaml'));
const swaggerJsonDocument = require('../swagger.json');
// Parse JSON request bodies
app.use(express.json());

// Serve static files from the public directory
app.use(express.static(path.join(__dirname, '..', 'public')));

// Root path serves index.html for convenience
app.get('/', (req, res) => {
  res.sendFile(path.join(__dirname, '..', 'public', 'index.html'));
});

// Endpoint to expose config values for client
app.get('/config', (req, res) => {
  res.json({
    weatherEndpoint: '/weather/current',
    astrologyEndpoint: '/astrology/forecast',
    defaultCity: DEFAULT_CITY,
    defaultSign: DEFAULT_SIGN
  });
});

// Serve OpenAPI specification
app.get('/openapi', (req, res) => {
  res.sendFile(path.join(__dirname, '..', 'cimeika-api.yaml'));
});

// Interactive API docs
app.use('/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(swaggerJsonDocument));

// Chat completion endpoint (mock)
app.post('/chat/completion', (req, res) => {
  const { prompt } = req.body || {};
  if (!prompt) return res.status(400).json({ error: 'prompt required' });
  res.json({
    id: 'chat1',
    object: 'text_completion',
    created: Date.now(),
    model: 'mock',
    choices: [{ text: `Echo: ${prompt}`, index: 0, logprobs: null, finish_reason: 'length' }]
  });
});


// Hugging Face completion endpoint
app.post('/ai/huggingface/completion', requireHfToken, huggingfaceCompletion);

// Hugging Face Space proxy endpoint
app.post('/ai/hf-space/completion', hfSpaceProxy);

// Mount feature routers
app.use('/auth', authRouter);
app.use('/components', componentsRouter);
app.use('/data', dataRouter);
app.use('/api/v1', apiV1Router);
app.use('/', realtimeRouter);

app.use((req, res) => res.status(404).json({ error: 'not found' }));

module.exports = app;