# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cimeika API",
    "description": "API для інтеграції GPT у існуючий інтерфейс чату та управління системою поєднаних компонентів",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://api.cimeika.com/v1",
      "description": "Основний сервер"
    }
  ],
  "paths": {
...
    "/chat/completion": {
      "post": {
        "operationId": "createChatCompletion",
        "summary": "Створення запиту до GPT для відповіді у чаті",
        "description": "Endpoint для створення запиту до GPT для отримання відповіді у чаті",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Модель GPT для використання",
                    "example": "text-davinci-003"
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Запит для моделі GPT",
                    "example": "Напиши доповідь про сучасні технології штучного інтелекту."
                  },
                  "max_tokens": {
                    "type": "integer",
                    "description": "Максимальна кількість токенів у відповіді",
                    "example": 100
                  },
                  "temperature": {
                    "type": "number",
                    "description": "Температура вибірки",
                    "example": 0.7
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Відповідь від GPT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "cmpl-5t6fN9FPjsMxlm8Zy1dJ1Z7P"
                    },
                    "object": {
                      "type": "string",
                      "example": "text_completion"
                    },
                    "created": {
                      "type": "integer",
                      "example": 1591685959
                    },
                    "model": {
                      "type": "string",
                      "example": "text-davinci-003"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "text": {
                            "type": "string",
                            "example": "Сучасні технології штучного інтелекту включають глибинне навчання, нейронні мережі та обробку природної мови..."
                          },
                          "index": {
                            "type": "integer",
                            "example": 0
                          },
                          "logprobs": {
                            "type": "object",
                            "nullable": true
                          },
                          "finish_reason": {
                            "type": "string",
                            "example": "length"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/huggingface/completion": {
      "post": {
        "operationId": "createHuggingFaceCompletion",
        "summary": "Створення запиту до Hugging Face GPT для відповіді у чаті",
        "description": "Endpoint для створення запиту до Hugging Face GPT для отримання відповіді у чаті. Потребує змінної середовища `HUGGINGFACE_TOKEN`; повертає 503, якщо токен не встановлено.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Модель GPT на Hugging Face для використання",
                    "example": "gpt-neo-2.7B"
                  },
                  "prompt": {
                    "type": "string",
                    "description": "Запит для моделі GPT",
                    "example": "Напиши доповідь про сучасні технології штучного інтелекту."
                  },
                  "max_tokens": {
                    "type": "integer",
                    "description": "Максимальна кількість токенів у відповіді",
                    "example": 150
                  },
                  "temperature": {
                    "type": "number",
                    "description": "Температура вибірки",
                    "example": 0.6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Відповідь від Hugging Face GPT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "hf-5t6fN9FPjsMxlm8Zy1dJ1Z7P"
                    },
                    "object": {
                      "type": "string",
                      "example": "text_completion"
                    },
                    "created": {
                      "type": "integer",
                      "example": 1591685959
                    },
                    "model": {
                      "type": "string",
                      "example": "gpt-neo-2.7B"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "text": {
                            "type": "string",
                            "example": "Сучасні технології штучного інтелекту включають глибинне навчання, нейронні мережі та обробку природної мови..."
                          },
                          "index": {
                            "type": "integer",
                            "example": 0
                          },
                          "logprobs": {
                            "type": "object",
                            "nullable": true
                          },
                          "finish_reason": {
                            "type": "string",
                            "example": "length"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/hf-space/completion": {
      "post": {
        "operationId": "proxyHuggingFaceSpace",
        "summary": "Проксі до Hugging Face Space \"cimeika-api\"",
        "description": "Надсилає запит у розгорнутий Space `https://ihorog-cimeika-api.hf.space/chat/completion` або на адресу, передану в полі `spaceUrl`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Запит для віддаленого Space",
                    "example": "Який сьогодні настрій у Cimeika?"
                  },
                  "spaceUrl": {
                    "type": "string",
                    "description": "Необов'язково. Перевизначення URL Space; за замовчуванням використовується змінна середовища `HF_SPACE_URL`.",
                    "example": "https://ihorog-cimeika-api.hf.space"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Відповідь від віддаленого Space",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextCompletionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Некоректний запит"
          },
          "502": {
            "description": "Space недоступний"
          }
        }
      }
    },
...
    "/components": {
      "get": {
        "operationId": "getComponents",
        "summary": "Отримати список всіх компонентів",
        "responses": {
          "200": {
            "description": "Список компонентів успішно отриманий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Component"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createComponent",
        "summary": "Створити новий компонент",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Component"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Компонент успішно створений",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Component"
                }
              }
            }
          }
        }
      }
    },
...
    "/weather/current": {
      "get": {
        "operationId": "getCurrentWeather",
        "summary": "Отримати поточні погодні дані",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "description": "Назва міста. Якщо параметр не вказано або порожній, сервер за замовчуванням використовує London.",
            "schema": {
              "type": "string",
              "example": "London",
              "default": "London",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Поточні погодні дані успішно отримані",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "city": {
                      "type": "string",
                      "description": "Назва міста",
                      "example": "London"
                    },
                    "weather": {
                      "type": "string",
                      "description": "Опис погоди",
                      "example": "clear sky"
                    },
                    "temperature": {
                      "type": "number",
                      "description": "Температура в градусах Цельсія",
                      "example": 20.5
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/astrology/forecast": {
      "get": {
        "operationId": "getAstrologicalForecast",
        "summary": "Отримати астрологічний прогноз",
        "parameters": [
          {
            "name": "sign",
            "in": "query",
            "required": false,
            "description": "Знак зодіаку. Якщо параметр не вказано або порожній, сервер за замовчуванням використовує aries.",
            "schema": {
              "type": "string",
              "example": "aries",
              "default": "aries",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Астрологічний прогноз успішно отриманий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sign": {
                      "type": "string",
                      "description": "Знак зодіаку",
                      "example": "aries"
                    },
                    "forecast": {
                      "type": "string",
                      "description": "Опис астрологічного прогнозу",
                      "example": "Today is a good day for new beginnings."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Component": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Унікальний ідентифікатор компоненту",
            "example": "component-12345"
          },
          "name": {
            "type": "string",
            "description": "Назва компоненту",
            "example": "Компонент 1"
          },
          "type": {
            "type": "string",
            "description": "Тип компоненту",
            "example": "базовий"
          },
          "attributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attribute"
            }
          }
        }
      },
      "Attribute": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Назва атрибуту",
            "example": "атрибут-1"
          },
          "value": {
            "type": "string",
            "description": "Значення атрибуту",
            "example": "значення-1"
          }
        }
      },
      "TextCompletionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Унікальний ідентифікатор відповіді"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "description": "Час створення відповіді"
          },
          "model": {
            "type": "string",
            "description": "Модель Hugging Face Space, яка використана"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "description": "Згенерований текст"
                },
                "index": {
                  "type": "integer",
                  "description": "Порядковий номер варіанту"
                },
                "finish_reason": {
                  "type": "string",
                  "description": "Причина завершення генерації"
                }
              }
            }
          }
        }
      },
...
    }
  }
}